Lassi module file renamed
