+++ my_pyscf/mcscf/addons.py
@@ -197,7 +197,7 @@ def get_h1e_zipped_fcisolver (fcisolver):
     return h1ezipped_fcisolver
 
 def las2cas_civec (las):
-    from mrh.my_pyscf.lassi.lassi_op_o0 import ci_outer_product
+    from mrh.my_pyscf.lassi.op_o0 import ci_outer_product
     norb_f = las.ncas_sub
     nelec_fr = [[_unpack_nelec (fcibox._get_nelec (solver, nelecas)) for solver in fcibox.fcisolvers] for fcibox, nelecas in zip (las.fciboxes, las.nelecas_sub)]
     ci, nelec = ci_outer_product (las.ci, norb_f, nelec_fr)
