+++ my_pyscf/mcscf/lassi_dms.py
@@ -1,3 +1,3 @@
-from mrh.my_pyscf.lassi.lassi_dms import *
+from mrh.my_pyscf.lassi.dms import *
 from mrh.util.io import lassi_dir_warn
 lassi_dir_warn ()
