+++ my_pyscf/mcscf/lassi_op_o1.py
@@ -1,3 +1,3 @@
-from mrh.my_pyscf.lassi.lassi_op_o1 import *
+from mrh.my_pyscf.lassi.op_o1 import *
 from mrh.util.io import lassi_dir_warn
 lassi_dir_warn ()
