+++ tests/lasscf/test_lassi_soc.py
@@ -7,7 +7,7 @@ from c2h6n4_struct import structure as struct
 from mrh.my_pyscf.fci import csf_solver
 from mrh.my_pyscf.mcscf import lassi_dms
 from mrh.my_pyscf.mcscf.soc_int import compute_hso, amfi_dm
-from mrh.my_pyscf.lassi.lassi_op_o0 import ci_outer_product
+from mrh.my_pyscf.lassi.op_o0 import ci_outer_product
 from mrh.my_pyscf.mcscf.lasscf_o0 import LASSCF
 from mrh.my_pyscf.lassi.lassi import make_stdm12s, roots_make_rdm12s, ham_2q
 import itertools
